refactor: extract Sudoku generation logic to service layer
SudokuGenerator Service:
- Extract all generation algorithms from useSudokuBoard hook
- Implement generateSolvedBoard() - creates complete valid board
- Implement generatePuzzleBoard() - creates puzzle with difficulty
- Implement validateCellPlacement() - validates player moves
- Use backtracking algorithm with randomization

useSudokuBoard Hook:
- Import and use SudokuGenerator service
- Simplify isValidPlacement to use service function
