+++ src/hooks/useSudokuBoard.js
@@ -1,115 +1,40 @@
 ﻿import { useState, useCallback } from "react";
-import { SUDOKU, DIFFICULTY_SETTINGS } from "../constants";
+import { SUDOKU } from "../constants";
+import { 
+  generatePuzzleBoard, 
+  validateCellPlacement 
+} from "../services/sudoku";
 
-function shuffleArray(array) {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-}
-
-function isValidPlacementInArray(board, row, col, num) {
-  for (let c = 0; c < SUDOKU.GRID_SIZE; c++) {
-    if (board[row][c] === num) return false;
-  }
-
-  for (let r = 0; r < SUDOKU.GRID_SIZE; r++) {
-    if (board[r][col] === num) return false;
-  }
-
-  const startRow = Math.floor(row / SUDOKU.SUBGRID_SIZE) * SUDOKU.SUBGRID_SIZE;
-  const startCol = Math.floor(col / SUDOKU.SUBGRID_SIZE) * SUDOKU.SUBGRID_SIZE;
-
-  for (let r = startRow; r < startRow + SUDOKU.SUBGRID_SIZE; r++) {
-    for (let c = startCol; c < startCol + SUDOKU.SUBGRID_SIZE; c++) {
-      if (board[r][c] === num) return false;
-    }
-  }
-
-  return true;
-}
-
-function fillBoard(board) {
-  for (let row = 0; row < SUDOKU.GRID_SIZE; row++) {
-    for (let col = 0; col < SUDOKU.GRID_SIZE; col++) {
-      if (board[row][col] === 0) {
-        const numbers = [1, 2, 3, 4, 5, 6, 7, 8, 9];
-        shuffleArray(numbers);
-
-        for (const num of numbers) {
-          if (isValidPlacementInArray(board, row, col, num)) {
-            board[row][col] = num;
-
-            if (fillBoard(board)) {
-              return true;
-            }
-
-            board[row][col] = 0;
-          }
-        }
-        return false;
-      }
-    }
-  }
-  return true;
-}
-
-function generateSolvedBoard() {
-  const board = Array(SUDOKU.GRID_SIZE)
-    .fill(null)
-    .map(() => Array(SUDOKU.GRID_SIZE).fill(0));
-
-  fillBoard(board);
-  return board;
-}
-
-function generateInitialBoard(difficultyLevel) {
-  const filledCells = DIFFICULTY_SETTINGS[difficultyLevel]?.filledCells || 30;
-  const solvedBoard = generateSolvedBoard();
-
-  const newBoard = [];
-  const totalCells = SUDOKU.GRID_SIZE * SUDOKU.GRID_SIZE;
-  const cellsToRemove = totalCells - filledCells;
-
-  const positions = [];
-  for (let row = 0; row < SUDOKU.GRID_SIZE; row++) {
-    for (let col = 0; col < SUDOKU.GRID_SIZE; col++) {
-      positions.push({ row, col });
-    }
-  }
-
-  shuffleArray(positions);
-
-  const cellsToRemoveSet = new Set();
-  for (let i = 0; i < cellsToRemove; i++) {
-    const pos = positions[i];
-    cellsToRemoveSet.add(`${pos.row}-${pos.col}`);
-  }
-
-  for (let row = 0; row < SUDOKU.GRID_SIZE; row++) {
-    const rowCells = [];
-    for (let col = 0; col < SUDOKU.GRID_SIZE; col++) {
-      const key = `${row}-${col}`;
-      const isPreFilled = !cellsToRemoveSet.has(key);
-
-      rowCells.push({
-        id: `cell-${row}-${col}`,
-        value: isPreFilled ? solvedBoard[row][col] : SUDOKU.EMPTY_CELL,
-        isEditable: !isPreFilled,
-        rowIndex: row,
-        colIndex: col,
-      });
-    }
-    newBoard.push(rowCells);
-  }
-  return newBoard;
-}
-
+/**
+ * Custom hook for managing Sudoku board state
+ * 
+ * This hook manages the game board state and provides methods to:
+ * - Update cells
+ * - Check board completion
+ * - Validate cell placements
+ * - Generate new games
+ * - Restore saved games
+ * 
+ * Note: Board generation logic has been moved to services/sudoku/SudokuGenerator
+ * 
+ * @param {string} difficulty - Game difficulty level
+ * @param {Object[][]|null} initialBoard - Optional initial board state for restoring
+ * @returns {Object} Board state and manipulation methods
+ * 
+ * @example
+ * const { board, updateCell, isBoardComplete, newGame } = useSudokuBoard('medium');
+ */
 const useSudokuBoard = (difficulty = "medium", initialBoard = null) => {
   const [board, setBoard] = useState(
-    () => initialBoard || generateInitialBoard(difficulty)
+    () => initialBoard || generatePuzzleBoard(difficulty)
   );
 
+  /**
+   * Update a cell value in the board
+   * @param {number} rowIndex - Row index (0-8)
+   * @param {number} colIndex - Column index (0-8)
+   * @param {number|string} value - New cell value
+   */
   const updateCell = useCallback((rowIndex, colIndex, value) => {
     setBoard((prevBoard) => {
       const newBoard = prevBoard.map((row) => row.map((cell) => ({ ...cell })));
@@ -122,65 +47,47 @@ const useSudokuBoard = (difficulty = "medium", initialBoard = null) => {
     });
   }, []);
 
+  /**
+   * Check if the board is completely filled
+   * @returns {boolean} True if all cells have values
+   */
   const isBoardComplete = useCallback(() => {
     return board.every((row) =>
       row.every((cell) => cell.value !== SUDOKU.EMPTY_CELL)
     );
   }, [board]);
 
+  /**
+   * Check if a value can be validly placed at a position
+   * Uses the validateCellPlacement service for validation logic
+   * @param {number} rowIndex - Row index
+   * @param {number} colIndex - Column index  
+   * @param {number|string} value - Value to validate
+   * @returns {boolean} True if placement is valid
+   */
   const isValidPlacement = useCallback(
     (rowIndex, colIndex, value) => {
-      if (!value || value === SUDOKU.EMPTY_CELL) return true;
-
-      for (let col = 0; col < SUDOKU.GRID_SIZE; col++) {
-        if (col !== colIndex && board[rowIndex][col].value === value) {
-          return false;
-        }
-      }
-
-      for (let row = 0; row < SUDOKU.GRID_SIZE; row++) {
-        if (row !== rowIndex && board[row][colIndex].value === value) {
-          return false;
-        }
-      }
-
-      const subgridRowStart =
-        Math.floor(rowIndex / SUDOKU.SUBGRID_SIZE) * SUDOKU.SUBGRID_SIZE;
-      const subgridColStart =
-        Math.floor(colIndex / SUDOKU.SUBGRID_SIZE) * SUDOKU.SUBGRID_SIZE;
-
-      for (
-        let row = subgridRowStart;
-        row < subgridRowStart + SUDOKU.SUBGRID_SIZE;
-        row++
-      ) {
-        for (
-          let col = subgridColStart;
-          col < subgridColStart + SUDOKU.SUBGRID_SIZE;
-          col++
-        ) {
-          if (
-            (row !== rowIndex || col !== colIndex) &&
-            board[row][col].value === value
-          ) {
-            return false;
-          }
-        }
-      }
-
-      return true;
+      return validateCellPlacement(board, rowIndex, colIndex, value);
     },
     [board]
   );
 
+  /**
+   * Reset the board with a new puzzle
+   * @param {string} newDifficulty - Optional new difficulty level
+   */
   const resetBoard = useCallback(
     (newDifficulty) => {
-      const newBoard = generateInitialBoard(newDifficulty || difficulty);
+      const newBoard = generatePuzzleBoard(newDifficulty || difficulty);
       setBoard(newBoard);
     },
     [difficulty]
   );
 
+  /**
+   * Start a new game with optional different difficulty
+   * @param {string} newDifficulty - Optional new difficulty level
+   */
   const newGame = useCallback(
     (newDifficulty) => {
       resetBoard(newDifficulty);
@@ -188,6 +95,10 @@ const useSudokuBoard = (difficulty = "medium", initialBoard = null) => {
     [resetBoard]
   );
 
+  /**
+   * Restore a previously saved board state
+   * @param {Object[][]} savedBoard - Saved board to restore
+   */
   const restoreBoard = useCallback((savedBoard) => {
     if (savedBoard) {
       setBoard(savedBoard);
