+++ src/services/sudoku/index.js
@@ -1,0 +1,7 @@
+export { 
+  generateSolvedBoard, 
+  generatePuzzleBoard, 
+  validateCellPlacement 
+} from './SudokuGenerator';
+
+export { default as SudokuGenerator } from './SudokuGenerator';
